Thêm api lấy job đã huấn luyện + phân trang
